business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetTuningStep.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetTuningStep implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+N ");
        messageString.append(command.getParameter("Tuning Step"));
        messageString.append("\n");
        return messageString;
    }
}
